requirements.txt
# Génération de documents LOI
python-docx>=1.1.0
openpyxl>=3.1.2
streamlit>=1.31.0

# Utilitaires
python-dateutil>=2.8.2

# API INPI
python-dotenv>=1.0.0
requests>=2.31.0
ratelimit>=2.2.1
